+++ affiliation_parser/parser/__init__.py
@@ -1,6 +1,6 @@
-from .check_country import check_country
-from .append_institution_city import append_institution_city
+# from .check_country import check_country
+# from .append_institution_city import append_institution_city
 from .clean_text import clean_text
-from .parse_email import parse_email
-from .parse_location import parse_location
-from .parse_zipcode import parse_zipcode
+# from .parse_email import parse_email
+# from .parse_location import parse_location
+# from .parse_zipcode import parse_zipcode
